Removed deprecated methods from file-list
