+++ lib/file-list.js
@@ -3,11 +3,8 @@
  * ========
  */
 var fs = require('fs');
-var vowFs = require('./fs/async-fs');
-var Vow = require('vow');
 var inherit = require('inherit');
 var path = require('path');
-var deprecate = require('./utils/deprecate');
 
 /**
  * FileList — класс для работы со списком файлов.
@@ -84,119 +81,11 @@ module.exports = inherit({
     },
 
     /**
-     * Синхронно загружает список файлов из директории.
-     * @param {String} dirname Абсолютный путь к директории.
-     * @param {Boolean} recursive Рекурсивная загрузка.
-     *
-     * Метод будет удалён в версии 1.0.0.
-     * @deprecated
-     */
-    loadFromDirSync: function (dirname, recursive) {
-        deprecate({method: 'loadFromDirSync', module: 'file-list', since: 'v1.0.0'});
-
-        var files = [];
-        var _this = this;
-        filterFiles(fs.readdirSync(dirname)).forEach(function (filename) {
-            var fullname = path.join(dirname, filename);
-            var stat = fs.statSync(fullname);
-            if (stat.isFile()) {
-                files.push({
-                    name: filename,
-                    fullname: fullname,
-                    suffix: getSuffix(filename),
-                    mtime: stat.mtime.getTime()
-                });
-            } else if (stat.isDirectory()) {
-                if (recursive) {
-                    _this.loadFromDirSync(fullname, recursive);
-                }
-            }
-        });
-        this.addFiles(files);
-    },
-
-    /**
-     * Асинхронно загружает список файлов из директории.
-     * @param {String} dirname Абсолютный путь к директории.
-     * @param {Boolean} recursive Рекурсивная загрузка.
-     * @returns {Promise}
-     *
-     * Метод будет удалён в версии 1.0.0.
-     * @deprecated
-     */
-    loadFromDir: function (dirname, recursive) {
-        deprecate({method: 'loadFromDir', module: 'file-list', since: 'v1.0.0'});
-
-        var _this = this;
-        return this._loadFromDir(dirname, recursive).then(function (files) {
-            _this.addFiles(files);
-        });
-    },
-
-    /**
-     * Асинхронно загружает список файлов из директории.
-     * @param {String} dirname
-     * @param {Boolean} recursive
-     * @returns {Promise}
-     * @private
-     *
-     * Метод будет удалён в версии 1.0.0.
-     * @deprecated
-     */
-    _loadFromDir: function (dirname, recursive) {
-        var _this = this;
-        return vowFs.listDir(dirname).then(function (filenames) {
-            return Vow.all(filenames.map(function (filename) {
-                var fullname = path.join(dirname, filename);
-                return vowFs.stat(fullname).then(function (stat) {
-                    if (stat.isFile()) {
-                        return [{
-                            name: filename,
-                            fullname: fullname,
-                            suffix: getSuffix(filename),
-                            mtime: stat.mtime.getTime()
-                        }];
-                    } else if (stat.isDirectory()) {
-                        if (recursive) {
-                            return _this._loadFromDir(fullname, recursive);
-                        }
-                    }
-                    return [];
-                });
-            }));
-        }).then(function (fileLists) {
-            return [].concat.apply([], fileLists);
-        });
-    },
-
-    /**
-     * Разбирает имя файла, извлекая данные BEM-сущности.
-     * Формат ответа:
-     * {
-     *   filenameWithoutSuffix: <имя файла без суффикса>,
-     *   suffix: <суффикс>,
-     *   bem: {
-     *       block: ...,
-     *       elem: ...,
-     *       modName: ...,
-     *       modVal: ...
-     *   },
-     *   bemdecl: <bemdecl-формат>
-     * };
-     * @returns {Object}
-     *
-     * Метод будет удалён в версии 1.0.0.
-     * @deprecated
-     */
-    parseFilename: parseFilename,
-
-    /**
      * Возвращает информацию о файле в виде:
      * { fullname: <абсолютный путь к файлу>, name: <имя файла>, suffix: <суффикс>, mtime: <время изменения> }
      */
     getFileInfo: getFileInfo
 }, {
-    parseFilename: parseFilename,
     getFileInfo: getFileInfo
 });
 
@@ -211,57 +100,4 @@ function getFileInfo(filename) {
         mtime: stat.mtime.getTime(),
         isDirectory: stat.isDirectory()
     };
-}
-
-function parseFilename(filename) {
-    deprecate({method: 'parseFilename', module: 'file-list', since: 'v1.0.0'});
-
-    var filenameParts = filename.split('.');
-    var filenameWithoutSuffix = filenameParts[0];
-    var suffix = filenameParts.slice(1).join('.');
-    var bem = {};
-    var modParts;
-    if (~filenameWithoutSuffix.indexOf('__')) {
-        var blockElemParts = filenameWithoutSuffix.split('__');
-        bem.block = blockElemParts[0];
-        var elemParts = blockElemParts[1].split('_');
-        bem.elem = elemParts[0];
-        modParts = elemParts.slice(1);
-    } else {
-        var blockParts = filenameWithoutSuffix.split('_');
-        bem.block = blockParts[0];
-        modParts = blockParts.slice(1);
-    }
-    if (modParts.length) {
-        bem.modName = modParts[0];
-        bem.modVal = modParts[1] || '';
-    }
-    var bemdecl = {
-        name: bem.block
-    };
-    if (bem.elem) {
-        if (bem.modName) {
-            bemdecl.elems = [{ name: bem.elem, mods: [{ name: bem.modName, vals: [ {name: bem.modVal} ] }]}];
-        } else {
-            bemdecl.elems = [{ name: bem.elem }];
-        }
-    } else if (bem.modName) {
-        bemdecl.mods = [{ name: bem.modName, vals: [ {name: bem.modVal} ] }];
-    }
-    return {
-        filenameWithoutSuffix: filenameWithoutSuffix,
-        suffix: suffix,
-        bem: bem,
-        bemdecl: bemdecl
-    };
-}
-
-function filterFiles(filenames) {
-    return filenames.filter(function (filename) {
-        return filename.charAt(0) !== '.';
-    });
-}
-
-function getSuffix(filename) {
-    return filename.split('.').slice(1).join('.');
 }
