+++ test/lib/build-flow.js
@@ -1015,7 +1015,7 @@ describe('build-flow', function () {
 
             var list = new FileList();
 
-            list.loadFromDirSync(dir);
+            list.addFiles(getFiles(dir));
             node.provideTechData('?.files', list);
 
             return node.runTechAndGetContent(Tech)
@@ -1765,7 +1765,7 @@ describe('build-flow', function () {
                         .builder(function () {})
                         .createTech();
 
-                    fileList.loadFromDirSync(blocksDirname);
+                    fileList.addFiles(getFiles(blocksDirname));
                     bundle.provideTechData('?.files', fileList);
 
                     return bundle.runTech(Tech)
@@ -1793,7 +1793,7 @@ describe('build-flow', function () {
                         })
                         .createTech();
 
-                    fileList.loadFromDirSync(blocksDirname);
+                    fileList.addFiles(getFiles(blocksDirname));
                     bundle.provideTechData('?.files', fileList);
 
                     return bundle.runTech(Tech)
@@ -1820,7 +1820,7 @@ describe('build-flow', function () {
                         })
                         .createTech();
 
-                    fileList.loadFromDirSync(blocksDirname);
+                    fileList.addFiles(getFiles(blocksDirname));
                     bundle.provideTechData('?.files', fileList);
 
                     return bundle.runTech(Tech)
@@ -1828,7 +1828,7 @@ describe('build-flow', function () {
                             helper.change(blockFilename);
 
                             fileList = new FileList();
-                            fileList.loadFromDirSync(blocksDirname);
+                            fileList.addFiles(getFiles(blocksDirname));
                             bundle.provideTechData('?.files', fileList);
 
                             return bundle.runTech(Tech);
@@ -2113,4 +2113,31 @@ function init(node, Tech, opts) {
     tech.init(node);
 
     return tech;
+}
+
+function getFiles(dirname) {
+    var files = [];
+    filterFiles(fs.readdirSync(dirname)).forEach(function (filename) {
+        var fullname = path.join(dirname, filename);
+        var stat = fs.statSync(fullname);
+        if (stat.isFile()) {
+            files.push({
+                name: filename,
+                fullname: fullname,
+                suffix: getSuffix(filename),
+                mtime: stat.mtime.getTime()
+            });
+        }
+    });
+    return files;
+}
+
+function filterFiles(filenames) {
+    return filenames.filter(function (filename) {
+        return filename.charAt(0) !== '.';
+    });
+}
+
+function getSuffix(filename) {
+    return filename.split('.').slice(1).join('.');
 }
